+++ config/webpack.config.js
@@ -4,20 +4,16 @@
  * SPDX-License-Identifier: LicenseRef-Nordic-4-Clause
  */
 
-const webpack = require('webpack');
-const path = require('path');
 const fs = require('fs');
-const ESLintPlugin = require('eslint-webpack-plugin');
+const path = require('path');
 
+// eslint-disable-next-line import/no-dynamic-require
 const { dependencies } = require(path.join(process.cwd(), 'package.json'));
 
 const appDirectory = fs.realpathSync(process.cwd());
-const nodeEnv = process.env.NODE_ENV || 'development';
-const isProd = nodeEnv === 'production';
 
-function createExternals() {
-    // Libs provided by nRF Connect at runtime
-    const coreLibs = [
+const externals = Object.fromEntries(
+    [
         'react',
         'react-dom',
         'react-redux',
@@ -29,98 +25,71 @@ function createExternals() {
         'pc-ble-driver-js',
         '@nordicsemiconductor/nrf-device-lib-js',
         'osx-temperature-sensor',
-    ];
+        ...Object.keys(dependencies),
+    ].map(lib => [lib, lib])
+);
 
-    // Libs provided by the app at runtime
-    const appLibs = Object.keys(dependencies);
+const entry = [
+    './src/index.jsx',
+    './lib/index.jsx',
+    './index.jsx',
+    './src/index.tsx',
+].find(fs.existsSync);
 
-    return coreLibs
-        .concat(appLibs)
-        .reduce((prev, lib) => Object.assign(prev, { [lib]: lib }), {});
-}
+module.exports = (_, argv) => {
+    const mode = argv.mode ?? 'production';
+    const isProd = mode === 'production';
 
-const eslintConfig = () => {
-    try {
-        return require.resolve('../../../.eslintrc');
-    } catch (err) {
-        return require.resolve('./eslintrc.json');
-    }
-};
-
-function findEntryPoint() {
-    const files = [
-        './src/index.jsx',
-        './lib/index.jsx',
-        './index.jsx',
-        './src/index.tsx',
-    ];
-    while (files.length) {
-        const file = files.shift();
-        if (fs.existsSync(file)) {
-            return file;
-        }
-    }
-    return undefined;
-}
-
-module.exports = {
-    mode: nodeEnv,
-    devtool: isProd ? 'source-map' : 'inline-cheap-source-map',
-    entry: findEntryPoint(),
-    output: {
-        path: path.join(appDirectory, 'dist'),
-        publicPath: './dist/',
-        filename: 'bundle.js',
-        libraryTarget: 'umd',
-    },
-    module: {
-        rules: [
-            {
-                test: /\.(jsx?|tsx?)$/,
-                use: [
-                    {
-                        loader: require.resolve('babel-loader'),
-                        options: {
-                            cacheDirectory: true,
-                            configFile:
-                                './node_modules/pc-nrfconnect-shared/config/babel.config.js',
+    return {
+        mode,
+        devtool: isProd ? 'source-map' : 'cheap-eval-source-map',
+        entry,
+        output: {
+            path: path.join(appDirectory, 'dist'),
+            publicPath: './dist/',
+            filename: 'bundle.js',
+            libraryTarget: 'umd',
+        },
+        module: {
+            rules: [
+                {
+                    test: /\.(jsx?|tsx?)$/,
+                    use: [
+                        {
+                            loader: require.resolve('babel-loader'),
+                            options: {
+                                cacheDirectory: true,
+                                configFile:
+                                    './node_modules/pc-nrfconnect-shared/config/babel.config.js',
+                            },
                         },
-                    },
-                ],
-                exclude: /node_modules\/(?!pc-nrfconnect-shared\/)/,
-            },
-            {
-                test: /\.scss|\.css$/,
-                use: [
-                    'style-loader',
-                    'css-loader',
-                    {
-                        loader: 'sass-loader',
-                        options: {
-                            // eslint-disable-next-line global-require
-                            implementation: require('sass'),
+                    ],
+                    exclude: /node_modules\/(?!pc-nrfconnect-shared\/)/,
+                },
+                {
+                    test: /\.scss|\.css$/,
+                    use: [
+                        'style-loader',
+                        'css-loader',
+                        {
+                            loader: 'sass-loader',
+                            options: {
+                                // eslint-disable-next-line global-require
+                                implementation: require('sass'),
+                            },
                         },
-                    },
-                ],
-            },
-            {
-                test: /\.(png|gif|ttf|eot|svg|woff(2)?)(\?[a-z0-9]+)?$/,
-                loader: require.resolve('url-loader'),
-            },
-        ],
-    },
-    resolve: {
-        extensions: ['.js', '.jsx', '.ts', '.tsx'],
-    },
-    plugins: [
-        new webpack.DefinePlugin({
-            'process.env.NODE_ENV': JSON.stringify(nodeEnv),
-        }),
-        new ESLintPlugin({
-            extensions: ['ts', 'tsx', 'js', 'jsx'],
-            overrideConfigFile: eslintConfig(),
-        }),
-    ],
-    target: 'electron-renderer',
-    externals: createExternals(),
+                    ],
+                },
+                {
+                    test: /\.(png|gif|ttf|eot|svg|woff(2)?)(\?[a-z0-9]+)?$/,
+                    loader: require.resolve('url-loader'),
+                },
+            ],
+        },
+        resolve: {
+            extensions: ['.js', '.jsx', '.ts', '.tsx'],
+        },
+        target: 'electron-renderer',
+        externals,
+    };
 };