+++ mocks/bleDriverMock.ts
@@ -6,6 +6,4 @@
 
 // Allows jest to test files that import pc-ble-driver-js.
 
-module.exports = {
-    AdapterFactory: { getInstance: () => {} },
-};
+export const AdapterFactory = { getInstance: () => {} };