+++ mocks/remoteMock.ts
@@ -6,9 +6,8 @@
 
 // Allows jest to test files that import the electron package.
 
-module.exports = {
-    require: jest.fn(),
-    getCurrentWindow: () => ({
-        reload: jest.fn(),
-    }),
-};
+export const require = jest.fn();
+export const dialog = jest.fn();
+export const getCurrentWindow = () => ({
+    reload: jest.fn(),
+});